Add app layout directory and fix login bug
