+++ app/routes/__app/profile.tsx
@@ -1,0 +1,3 @@
+export default function ProfilePage() {
+    return <h1>Profile</h1>
+}
