+++ app/routes/__app/users.tsx
@@ -1,8 +1,7 @@
-import { Form, json, useLoaderData, Outlet, Link, NavLink } from "remix";
+import { json, useLoaderData, Outlet, NavLink } from "remix";
 import type { LoaderFunction } from "remix";
 
 import { requireUserId } from "~/session.server";
-import { useUser } from "~/utils";
 import { getUsers } from "~/models/user.server";
 
 type LoaderData = {
@@ -17,25 +16,8 @@ export const loader: LoaderFunction = async ({ request }) => {
 
 export default function UsersPage() {
   const data = useLoaderData() as LoaderData;
-  const user = useUser();
-
   return (
     <div className="flex h-full min-h-screen flex-col">
-      <header className="flex items-center justify-between bg-slate-800 p-4 text-white">
-        <h1 className="text-3xl font-bold">
-          <Link to=".">Wildcards Watchlist</Link>
-        </h1>
-        <p>{user.email}</p>
-        <Form action="/logout" method="post">
-          <button
-            type="submit"
-            className="rounded bg-slate-600 py-2 px-4 text-blue-100 hover:bg-blue-500 active:bg-blue-600"
-          >
-            Logout
-          </button>
-        </Form>
-      </header>
-
       <main className="flex h-full bg-white">
         <div className="h-full w-80 border-r bg-gray-50">
           <ol>
